refactor: migrate console application using repository
